Moved saveCashRegister to live in the same folder as other processing methods
