+++ samples/snapshots/src/cashiers/processCashRegister/updateCashRegister.ts
@@ -1,6 +1,6 @@
 import { getEventStore } from '../../core/eventStore';
 import { CashRegisterEvent } from '../cashRegister';
-import { saveCashRegister } from '../saveCashRegister';
+import { saveCashRegister } from './saveCashRegister';
 import { STREAM_NOT_FOUND } from '../../core/eventStore/reading';
 import { getAndUpdate } from '../../core/eventStore/eventStoreDB/appending';
 import { readEventsFromSnapshotInSeparateStream } from '../../core/eventStore/eventStoreDB/reading/readFromSnapshotAndStream';
@@ -15,11 +15,8 @@ export function updateCashRegister<Command, TError = never>(
 ): Promise<boolean | STREAM_NOT_FOUND | TError | never> {
   return getAndUpdate<Command, CashRegisterEvent, TError>(
     getEventStore(),
-    (eventStore, streamName) =>
-      readEventsFromSnapshotInSeparateStream<CashRegisterEvent>(
-        eventStore,
-        streamName
-      ),
+    (...args) =>
+      readEventsFromSnapshotInSeparateStream<CashRegisterEvent>(...args),
     saveCashRegister,
     streamName,
     command,
